Pulled out parameters into separate module
